requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
cors==1.0.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
filelock==3.18.0
future==1.0.0
gevent==25.4.2
greenlet==3.2.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rpds-py==0.24.0
sqlparse==0.5.3
tldextract==5.3.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
zope.event==5.0
zope.interface==7.2
